Fix inconsistent my_find_package behaviour (glfw link issue on windows)
Original error: the glfw3.lib wasn't linked, because the GLFW_LIBRARIES variable
was empty (instead, only GLFW_LIBRARY was set).

Let's require that the first find_package call only uses cmake modules (and assume
that when/if such a module is available, it sets the plural variable as it should);
and if there is no module, continue with the manual `find_` calls, where we
use consistent variable naming.